Improved findaudio and findvideo
findaudio previously found only files that were in 5.1,
it now instead finds files that are *not* in stereo.

Similarly, findvideo finds files that are *not* H.265,
instead of finding files that are H.264.

There is now also a catch to check that files are mkv, mp4, or avi
as a text file would also say that it is not H.265